+++ R/theme-brain.R
@@ -28,12 +28,13 @@
 #' @examples
 #' library(ggplot2)
 #'
-#' ggplot() +
-#'   geom_brain(atlas = dk) +
+#' p <- ggplot() +
+#'   geom_brain(atlas = dk)
+#'
+#' p +
 #'   theme_brain()
 #'
-# ggplot() +
-#'   geom_brain(atlas = dk) +
+#' p +
 #'   theme_darkbrain()
 #'
 theme_brain = function(text.size=12,