+++ data-raw/make_data.R
@@ -1,4 +1,6 @@
-devtools::load_all(".")
+library(ggseg)
+library(dplyr)
+library(ggsegExtra)
 devtools::load_all("../ggsegExtra/")
 
 
@@ -34,38 +36,50 @@ aseg_n <- aseg %>%
 aseg_n$palette <- brain_pals$aseg
 
 
-aseg_n <- make_subcort_ggseg(output_dir = "data-raw/aseg",
-                             steps = 5:8,
-                             tolerance = 1,
-                             smoothness = 2,
-                             vertex_size_limits = c(15, NA))
-#
+aseg_n <- make_volumetric_ggseg(
+  label_file = "/Applications/freesurfer/7.2.0/subjects/fsaverage5/mri/aseg.mgz",
+  output_dir = "data-raw/aseg",
+  steps = 8,
+  tolerance = 1,
+  smoothness = 2,
+  vertex_size_limits = c(20, NA),
+  ncores = 16,
+  slices = dplyr::tribble(
+    ~x, ~y, ~z, ~view,
+    130, 130, 130, "axial",
+    128, 136, 100, "axial",
+    127, 135, 128, "coronal")
+)
+
+aseg2 <- aseg_n
 # # Do some data clean-up
-aseg_n$data <- aseg_n$data %>%
+aseg2$data <- aseg_n$data %>%
   filter(!is.na(region)) %>%
   mutate(region = gsub("cc ", "CC ", region),
          region = gsub("dc", " DC", region),
          region = ifelse(region == "cerebral cortex", NA, region)
   ) %>%
-  filter(!grepl("white|csf", region))
+  filter(!grepl("white|csf", region))  %>%
+  filter(side != "sagittal") |>
+  rbind(aseg$data |>
+          filter(side == "sagittal") |>
+          mutate(roi = NA), all = TRUE)
 
 
-names(aseg_n$palette) <- gsub("cc", "CC", names(aseg_n$palette))
-names(aseg_n$palette) <- gsub("dc", "DC", names(aseg_n$palette))
-aseg_n$palette <- aseg_n$palette[!grepl("white|csf|cerebral cortex", names(aseg_n$palette))]
+names(aseg2$palette) <- gsub("cc", "CC", names(aseg2$palette))
+names(aseg2$palette) <- gsub("dc", "DC", names(aseg2$palette))
+aseg2$palette <- aseg2$palette[!grepl("white|csf|cerebral cortex", names(aseg2$palette))]
 
-aseg_n %>%
-  ggseg(atlas = ., show.legend = TRUE,
-        colour = "black",
-        # position = "s",
-        mapping = aes(fill=region))
+ggseg(atlas = aseg2, show.legend = TRUE,
+      colour = "black",
+      mapping = aes(fill=region))
 
-plot(aseg_n)
+plot(aseg2)
 
 ggplot() +
-  geom_brain(atlas = aseg_n)
+  geom_brain(atlas = aseg2)
 
-aseg <- aseg_n
+aseg <- aseg2
 
 usethis::use_data(aseg,
                   internal = FALSE,